fix(admin) query by uuid username
Previously to this change, the admin api would only query by
alternative endpoint keys (like username or name) if the value isn't
a valid UUID, in which case it would consider it an id (primary key).
If an entity `name` or `username` was a valid uuid, the Admin API call
would return a `404 Not Found`.

A consequence of this fix is that GET requests on entity Admin API endpoints
will result in two requests if said entity has an endpoint key and the GET
was based on ID. First, it will attempt to query by username, then by ID.